+++ plugins/database/sqlite.h
@@ -52,7 +52,7 @@ class SQLiteDatabasePlugin : public DatabasePlugin {
   Status scan(const std::string& domain,
               std::vector<std::string>& results,
               const std::string& prefix,
-              size_t max) const override;
+              uint64_t max) const override;
 
  public:
   /// Database workflow: open and setup.
