+++ plugins/logger/buffered.h
@@ -49,7 +49,7 @@ inline void iterate(std::vector<std::string>& input,
 class BufferedLogForwarder : public InternalRunnable {
  protected:
   static const std::chrono::seconds kLogPeriod;
-  static const size_t kMaxLogLines;
+  static const uint64_t kMaxLogLines;
 
  protected:
   // These constructors are made available for subclasses to use, but
@@ -105,7 +105,7 @@ class BufferedLogForwarder : public InternalRunnable {
    *
    * @param s Results string to log
    */
-  Status logString(const std::string& s, size_t time = 0);
+  Status logString(const std::string& s, uint64_t time = 0);
 
   /**
    * @brief Log a vector of status lines
@@ -116,7 +116,7 @@ class BufferedLogForwarder : public InternalRunnable {
    *
    * @param log Vector of status lines to log
    */
-  Status logStatus(const std::vector<StatusLogLine>& log, size_t time = 0);
+  Status logStatus(const std::vector<StatusLogLine>& log, uint64_t time = 0);
 
  protected:
   /**
@@ -160,15 +160,15 @@ class BufferedLogForwarder : public InternalRunnable {
 
  protected:
   /// Generate a result index string to use with the backing store
-  std::string genResultIndex(size_t time = 0);
+  std::string genResultIndex(uint64_t time = 0);
 
   /// Generate a status index string to use with the backing store
-  std::string genStatusIndex(size_t time = 0);
+  std::string genStatusIndex(uint64_t time = 0);
 
  private:
   std::string genIndexPrefix(bool results);
 
-  std::string genIndex(bool results, size_t time = 0);
+  std::string genIndex(bool results, uint64_t time = 0);
 
   /**
    * @brief Add a database value while maintaining count
@@ -190,7 +190,7 @@ class BufferedLogForwarder : public InternalRunnable {
   std::chrono::seconds log_period_;
 
   /// Max number of logs to flush per check
-  size_t max_log_lines_;
+  uint64_t max_log_lines_;
 
   /**
    * @brief Name to use in index
@@ -206,7 +206,7 @@ class BufferedLogForwarder : public InternalRunnable {
   std::atomic<size_t> log_index_{0};
 
   /// Stores the count of buffered logs
-  size_t buffer_count_{0};
+  unsigned long long int buffer_count_{0};
 
   /// Protects the count of buffered logs
   RecursiveMutex count_mutex_;
